+++ src/main/java/nextstep/auth/application/GithubErrorDecoder.java
@@ -1,0 +1,12 @@
+package nextstep.auth.application;
+
+import feign.Response;
+import feign.codec.ErrorDecoder;
+import nextstep.auth.exception.AuthenticationException;
+
+public class GithubErrorDecoder implements ErrorDecoder {
+  @Override
+  public Exception decode(String methodKey, Response response) {
+    return new AuthenticationException();
+  }
+}
