+++ src/main/java/nextstep/auth/application/dto/GithubLoginRequest.java
@@ -1,0 +1,12 @@
+package nextstep.auth.application.dto;
+
+import lombok.Getter;
+
+@Getter
+public class GithubLoginRequest {
+  private final String code;
+
+  public GithubLoginRequest(String code) {
+    this.code = code;
+  }
+}
