+++ src/main/java/nextstep/auth/application/dto/TokenResponse.java
@@ -1,0 +1,16 @@
+package nextstep.auth.application.dto;
+
+import lombok.Getter;
+
+@Getter
+public class TokenResponse {
+  private final String accessToken;
+
+  public TokenResponse() {
+    this(null);
+  }
+
+  public TokenResponse(String accessToken) {
+    this.accessToken = accessToken;
+  }
+}
