+++ src/main/java/nextstep/auth/config/AuthConfig.java
@@ -1,0 +1,20 @@
+package nextstep.auth.config;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import nextstep.auth.application.JwtTokenProvider;
+import nextstep.auth.ui.AuthenticationPrincipalArgumentResolver;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class AuthConfig implements WebMvcConfigurer {
+  private final JwtTokenProvider jwtTokenProvider;
+
+  @Override
+  public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+    argumentResolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
+  }
+}
