+++ src/main/java/nextstep/auth/exception/AuthorizationException.java
@@ -1,0 +1,10 @@
+package nextstep.auth.exception;
+
+import nextstep.support.error.ApiException;
+import nextstep.support.error.ErrorCode;
+
+public class AuthorizationException extends ApiException {
+  public AuthorizationException() {
+    super(ErrorCode.FORBIDDEN);
+  }
+}
