+++ src/main/java/nextstep/auth/ui/AuthenticationPrincipal.java
@@ -1,0 +1,10 @@
+package nextstep.auth.ui;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthenticationPrincipal {}
