+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -3,6 +3,8 @@ package nextstep.favorite.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import nextstep.auth.domain.LoginMember;
+import nextstep.auth.exception.AuthorizationException;
 import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
@@ -10,9 +12,7 @@ import nextstep.favorite.domain.FavoriteRepository;
 import nextstep.favorite.exception.FavoriteNotFoundException;
 import nextstep.favorite.exception.FavoritePathNotFoundException;
 import nextstep.member.application.MemberService;
-import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
-import nextstep.member.exception.AuthorizationException;
 import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.application.dto.PathRequest;
 import nextstep.subway.path.domain.Path;
