--- src/main/java/nextstep/member/AuthConfig.java
@@ -1,20 +1,0 @@
-package nextstep.member;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import nextstep.member.application.JwtTokenProvider;
-import nextstep.member.ui.AuthenticationPrincipalArgumentResolver;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class AuthConfig implements WebMvcConfigurer {
-  private final JwtTokenProvider jwtTokenProvider;
-
-  @Override
-  public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
-    argumentResolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
-  }
-}
