--- src/main/java/nextstep/member/application/GithubProfileClient.java
@@ -1,15 +1,0 @@
-package nextstep.member.application;
-
-import lombok.RequiredArgsConstructor;
-import nextstep.member.application.dto.GithubProfileResponse;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class GithubProfileClient {
-  private final GithubProfileApi githubProfileApi;
-
-  public GithubProfileResponse getProfile(String accessToken) {
-    return githubProfileApi.getProfile("Bearer " + accessToken);
-  }
-}
