+++ src/main/java/nextstep/member/application/MemberService.java
@@ -1,9 +1,9 @@
 package nextstep.member.application;
 
 import lombok.RequiredArgsConstructor;
+import nextstep.auth.domain.LoginMember;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
