--- src/main/java/nextstep/member/application/dto/TokenResponse.java
@@ -1,16 +1,0 @@
-package nextstep.member.application.dto;
-
-import lombok.Getter;
-
-@Getter
-public class TokenResponse {
-  private final String accessToken;
-
-  public TokenResponse() {
-    this(null);
-  }
-
-  public TokenResponse(String accessToken) {
-    this.accessToken = accessToken;
-  }
-}
