--- src/main/java/nextstep/member/exception/AuthenticationException.java
@@ -1,10 +1,0 @@
-package nextstep.member.exception;
-
-import nextstep.support.error.ApiException;
-import nextstep.support.error.ErrorCode;
-
-public class AuthenticationException extends ApiException {
-  public AuthenticationException() {
-    super(ErrorCode.UNAUTHORIZED);
-  }
-}
