--- src/main/java/nextstep/member/exception/AuthorizationException.java
@@ -1,10 +1,0 @@
-package nextstep.member.exception;
-
-import nextstep.support.error.ApiException;
-import nextstep.support.error.ErrorCode;
-
-public class AuthorizationException extends ApiException {
-  public AuthorizationException() {
-    super(ErrorCode.FORBIDDEN);
-  }
-}
