--- src/main/java/nextstep/member/ui/AuthenticationPrincipal.java
@@ -1,10 +1,0 @@
-package nextstep.member.ui;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticationPrincipal {}
