+++ src/main/java/nextstep/member/ui/MemberController.java
@@ -2,10 +2,11 @@ package nextstep.member.ui;
 
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
+import nextstep.auth.domain.LoginMember;
+import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
